Added completed text

+++ contracts/levels/CoinFlip.sol
@@ -1,8 +1,8 @@
 pragma solidity ^0.4.18;
 
 contract CoinFlip {
-  uint public consecutiveWins;
-  uint lastHash;
+  uint256 public consecutiveWins;
+  uint256 lastHash;
   uint256 FACTOR = 57896044618658097711785492504343953926634992332820282019728792003956564819968;
   
   function CoinFlip() public {
